+++ src/QNEthernetFrame.h
@@ -177,18 +177,23 @@ class EthernetFrameClass final : public Stream, public internal::PrintfChecked {
   // This returns data() + 14.
   const uint8_t *payload() const;
 
-  // Sets the receive queue size. This will use a minimum of 1.
+  // Sets the receive queue capacity. This will use a minimum of 1.
   //
-  // If the new size is smaller than the number of elements in the queue then
-  // all the oldest frames that don't fit are dropped.
+  // If the new capacity is smaller than the number of elements in the queue
+  // then all the oldest frames that don't fit are dropped.
   //
   // This disables interrupts while changing the queue so as not to interfere
   // with the receive function if called from an ISR.
-  void setReceiveQueueSize(size_t size);
+  void setReceiveQueueCapacity(size_t capacity);
 
-  // Returns the current receive queue size.
-  size_t receiveQueueSize() const {
+  // Returns the receive queue capacity.
+  size_t receiveQueueCapacity() const {
     return inBuf_.size();
+  }
+
+  // Returns the number of frames currently in the receive queue.
+  size_t receiveQueueSize() const {
+    return inBufSize_;
   }
 
   // Clears any outgoing packet and the incoming queue.